refactor backend spec framework
refactor backend spec framework prt 2
