+++ backend/spec/spec_helper.rb
@@ -13,6 +13,7 @@ require_relative "../app/model/db"
 require_relative "json_record_spec_helper"
 require_relative "custom_matchers"
 
+
 Dir.glob(File.join(File.dirname(__FILE__), '../', '../', 'common', 'lib', "*.jar")).each do |file|
   require file
 end
@@ -78,7 +79,6 @@ module JSONModel
 
     extend Rack::Test::Methods
 
-
     def self.multipart_request(uri, params)
       Struct.new(:method, :path, :body).new("POST", uri, params)
     end
@@ -97,11 +97,6 @@ module JSONModel
 end
 
 
-# Note: This import is loading JSONModel into the Object class.  Pretty gross!
-# It would be nice if we could narrow the scope of this to just the tests.
-include JSONModel
-
-
 # Switch off notifications for the tests
 require_relative '../app/lib/notifications'
 class Notifications
@@ -118,8 +113,6 @@ end
 
 
 require_relative "../app/main"
-
-
 
 Log.quiet_please
 
@@ -141,72 +134,8 @@ class ArchivesSpaceService
   end
 
   use ExceptionPrintingMiddleware
-end
 
 
-def app
-  ArchivesSpaceService
-end
-
-require 'factory_girl'
-
-
-# FactoryGirl.definition_file_paths = [File.dirname(__FILE__)]
-# FactoryGirl.find_definitions
-require_relative 'factories'
-include FactoryGirl::Syntax::Methods
-
-
-def make_test_repo(code = "ARCHIVESSPACE", org_code = "test")
-  repo = create(:repo, {:repo_code => code, :org_code => org_code})
-
-  @repo_id = repo.id
-  @repo = JSONModel(:repository).uri_for(repo.id)
-
-  JSONModel::set_repository(@repo_id)
-  RequestContext.put(:repo_id, @repo_id)
-
-  @repo_id
-end
-
-
-def make_test_user(username, name = "A test user", source = "local")
-  create(:user, {:username => username, :name => name, :source => source})
-end
-
-
-def create_accession(opts = {})
-  Accession.create_from_json(build(:json_accession,
-                                   {:title => "Papers of Mark Triggs"}.merge(opts)),
-                             :repo_id => $repo_id)
-end
-
-
-def create_agent_person(opts = {})
-  AgentPerson.create_from_json(build(:json_agent_person, opts),
-                               :repo_id => $repo_id)
-end
-
-
-def create_event(opts = {})
-  Event.create_from_json(build(:json_event, opts),
-                         :repo_id => $repo_id)
-end
-
-
-def create_resource(opts = {})
-  Resource.create_from_json(build(:json_resource, opts), :repo_id => $repo_id)
-end
-
-
-
-def create_digital_object(opts = {})
-  DigitalObject.create_from_json(build(:json_digital_object, opts), :repo_id => $repo_id)
-end
-
-
-
-class ArchivesSpaceService
   def current_user
     Thread.current[:active_test_user] or raise "Unknown user"
   end
@@ -220,70 +149,36 @@ class ArchivesSpaceService
 end
 
 
-def create_nobody_user
-  user = create(:user, :username => 'nobody')
-
-  viewers = JSONModel(:group).all(:group_code => "repository-viewers").first
-  viewers.member_usernames = ['nobody']
-  viewers.save
-
-  user
+def app
+  ArchivesSpaceService
 end
 
 
-def as_test_user(username)
-  old_user = Thread.current[:active_test_user]
-  Thread.current[:active_test_user] = User.find(:username => username)
-  orig = RequestContext.get(:enforce_suppression)
-  old_username = RequestContext.get(:current_username)
-
-  begin
-    if RequestContext.active?
-      RequestContext.put(:enforce_suppression,
-                         !Thread.current[:active_test_user].can?(:manage_repository))
-      RequestContext.put(:current_username, username)
-    end
-
-    yield
-  ensure
-    RequestContext.put(:enforce_suppression, orig) if RequestContext.active?
-    RequestContext.put(:current_username, old_username) if RequestContext.active?
-    Thread.current[:active_test_user] = old_user
-  end
-end
-
-def as_anonymous_user
-  old_user = Thread.current[:active_test_user]
-  orig = RequestContext.get(:enforce_suppression)
-
-  Thread.current[:active_test_user] = AnonymousUser.new
-
-  begin
-    if RequestContext.active?
-      RequestContext.put(:enforce_suppression, true)
-    end
-
-    yield
-  ensure
-    RequestContext.put(:enforce_suppression, orig) if RequestContext.active?
-    Thread.current[:active_test_user] = old_user
-  end
-end
-
-
-DB.open(true) do
-  RequestContext.open do
-    create(:agent_corporate_entity)
-    create(:repo)
-    $default_repo = $repo_id
-  end
-end
+require_relative 'factories'
+require_relative "spec_helper_methods"
 
 
 RSpec.configure do |config|
   config.include Rack::Test::Methods
   config.include FactoryGirl::Syntax::Methods
+  config.include SpecHelperMethods
+  config.include JSONModel
   config.treat_symbols_as_metadata_keys_with_true_values = true
+
+  # inclusions not in effect here
+  config.before(:suite) do
+    DB.open(true) do
+      SpecHelperMethods.as_test_user("admin") do
+        RequestContext.open do
+          FactoryGirl.create(:agent_corporate_entity)
+          FactoryGirl.create(:repo)
+          $default_repo = $repo_id
+          $repo_record = JSONModel.JSONModel(:repository).find($repo_id)
+        end
+      end
+    end
+  end
+
 
 #  Roll back the database after each test
   config.around(:each) do |example|
